pyproject.toml
[tool.poetry]
name = "sparfify-data-pipelines"
version = "0.1.0"
description = ""
authors = ["HarrisonHenri <harrisonhenrisn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
apache-airflow = "^2.7.1"
boto3 = "^1.28.38"
python-decouple = "^3.8"
boto3-stubs-lite = {extras = ["ec2", "redshift-serverless", "iam", "s3"], version = "^1.28.38"}
boto3-stubs = "^1.28.38"
botocore = "^1.31.38"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
